doc: updated documentation site
* domains: defined "async" domain as a better fit than "condition" for
  all async testing assertions
* maintainers: documented how to add a go version guard assertion; fixed
  codegen tests & go.work undue upgrade
* other minor documentation additions
