+++ assert/assert_assertions.go
@@ -3417,7 +3417,7 @@ func YAMLEqT[EDoc, ADoc RText](t T, expected EDoc, actual ADoc, msgAndArgs ...an
 //
 // # Usage
 //
-//	actual := struct {
+//	expected := struct {
 //		A int `yaml:"a"`
 //	}{
 //		A: 10,
