+++ docs/doc-site/api/file.md
@@ -1,7 +1,7 @@
 ---
 title: "File"
 description: "Asserting OS Files"
-weight: 7
+weight: 8
 domains:
   - "file"
 keywords:
