+++ docs/doc-site/api/http.md
@@ -1,7 +1,7 @@
 ---
 title: "Http"
 description: "Asserting HTTP Response And Body"
-weight: 8
+weight: 9
 domains:
   - "http"
 keywords:
