+++ docs/doc-site/api/json.md
@@ -1,7 +1,7 @@
 ---
 title: "Json"
 description: "Asserting JSON Documents"
-weight: 9
+weight: 10
 domains:
   - "json"
 keywords:
