+++ docs/doc-site/api/string.md
@@ -1,7 +1,7 @@
 ---
 title: "String"
 description: "Asserting Strings"
-weight: 14
+weight: 15
 domains:
   - "string"
 keywords:
