+++ docs/doc-site/api/time.md
@@ -1,7 +1,7 @@
 ---
 title: "Time"
 description: "Asserting Times And Durations"
-weight: 16
+weight: 17
 domains:
   - "time"
 keywords:
