+++ docs/doc-site/project/maintainers/ARCHITECTURE.md
@@ -13,29 +13,49 @@ We want the maintenance of dozens of test assertions, times many variants, to re
 The maintenance flow is intended to require different activities and levels of understanding,
 depending on the complexity of a planned evolution.
 
+Not everything can be hassle-free, but the design should offer a relatively easy path to most common maintenance.
+
+After 9 months of all sorts of maintenance tasks, the diagram below gives a rather faithful representation of what
+it actually costs.
+
 {{< mermaid align="center" zoom="true" >}}
-journey
-    section Fixes & minor enhancements
-      internal/assertions:5: Knowledge of the functionality
-    section New dependencies
-      internal/assertions/enable/...:5: Understanding of the repo architecture
-      enable/...:5:  Understanding of the repo architecture
-    section API changes
-      regenerate code:5: No specific knowledge
-    section New constructs to support
-      code & doc generator:5: Knowledge of internals
+quadrantChart
+    title Change complexity vs Required knowledge
+    x-axis Minimal Knowledge --> In-Depth Understanding
+    y-axis Simple Change --> Complex Change
+
+    quadrant-3 Hassle-free
+    quadrant-2 Follow the tracks
+    quadrant-4 Should stay empty
+    quadrant-1 Generator work
+
+    Bug fixes: [0.2, 0.2]
+    Doc fixes: [0.25, 0.1]
+    Minor enhancements: [0.3, 0.4]
+    New assertions: [0.35, 0.23]
+    New dependencies: [0.20, 0.70]
+    New constructs: [0.78, 0.78]
+    Guarded assertions: [0.40, 0.58]
+    Package layout: [0.68, 0.62]
+
 {{< /mermaid >}}
+
+The bottom-right quadrant is empty on purpose: nothing here asks for in-depth knowledge of the
+generator to make a small change. A point landing there means the code needs reshaping.
 
 Most common maintenance tasks should not require much more than fixing/enhancing the code in `internal/assertions`.
 
-API changes need an extra code generation.
+Fixes and enhancements propagate naturally to the variants without the need to regenerate code.
 
-Dependency changes (adding new features that need extra dependencies) is a bit more involved, but still manageable.
+API changes need an extra code generation, but no specific knowledge of the generator itself.
 
-The code & doc generator should rapidly become a very stable component. The maintenance of the generator itself remains
+Dependency changes (adding new features that need extra dependencies) is a bit more involved, but still manageable:
+the pattern is regular, follow the tracks.
+
+The code & doc generator has now become a rather stable component. The maintenance of the generator itself remains
 an operation that requires an extended understanding of the internals of the project.
 
-Fixes and enhancements propagate naturally to the variants without the need to regenerate code.
+Example of recent updates that required such in-depth maintenance: adding support for build guards.
 
 ### The maths with assertion variants
 
@@ -48,10 +68,11 @@ Each of these variants produces another formatted variant, plus one "forward" va
 
 **For every "helper" function (not an assertion): 2 variants.**
 
-Generic assertions reach 8 variants only from go1.27, the first release that accepts type parameters
-on methods. Their 4 method variants are generated into files guarded by `//go:build go1.27`
-(`assert/assert_forward_go127.go`, `require/require_forward_go127.go`), so a go1.25 or go1.26 build
-drops them and keeps the 4 package-level variants. The counts below assume go1.27.
+> Generic assertions reach 8 variants only from go1.27, the first release that accepts type parameters on methods.
+>
+> Their 4 method variants are generated into files guarded by `//go:build go1.27`
+> (`assert/assert_forward_go127.go`, `require/require_forward_go127.go`), so a go1.26 build
+> drops them and keeps the 4 package-level variants. The counts below assume go1.27.
 
 
 All these variants make up several hundreds functions, which poses a challenge for maintenance and documentation.
@@ -124,8 +145,8 @@ Exceptions:
 
 The `enable/` package provides optional features that users can activate via blank imports:
 - `enable/stubs/` - Public stub APIs for enabling features (yaml, colors)
-- `enable/yaml/` - Activates YAML support via `import _ "github.com/go-openapi/testify/v2/enable/yaml"`
-- `enable/colors/` - Activates colorized output via `import _ "github.com/go-openapi/testify/v2/enable/colors"`
+- `enable/yaml/` - Activates YAML support via `import _ "github.com/go-openapi/testify/enable/yaml/v2"`
+- `enable/colors/` - Activates colorized output via `import _ "github.com/go-openapi/testify/enable/colors/v2"`
 
 These packages are not generated and allow optional dependencies to be isolated from the core library.
